+++ tests/e2e/cases/dynamic_require_static_numeric_parser/f7_5.ts
@@ -1,0 +1,1 @@
+module.exports.label = "f7_5";